Fixed errors after testing, completing

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: adiaz-lo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 16:17:54 by adiaz-lo          #+#    #+#             */
-/*   Updated: 2023/04/24 16:17:49 by adiaz-lo         ###   ########.fr       */
+/*   Updated: 2023/05/01 11:02:01 by adiaz-lo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	if (!haystack)
 		return (0);
